refactor(docker): change docker image output format to oci
The output format for Docker images has been changed from 'docker' to 'oci'. This change is made to leverage the `docker buildx imagetools create` command, which allows for more efficient loading of OCI archives into the Docker daemon.

This also includes the addition of the `docker/setup-buildx-action` step to ensure `buildx` is available for `imagetools`.